기기등록 완료

+++ src/main/java/com/ht/connected/home/backend/config/service/ZwaveClassKey.java
@@ -7,7 +7,7 @@ public class ZwaveClassKey {
     //0x52
 	public static byte NETWORK_MANAGEMENT_PROXY = 0x52;
 	public static byte NETWORK_MANAGEMENT_BASIC = 0x4D;
-	public static int NETWORK_MANAGEMENT_INCLUSION = 0x34;
+	public static byte NETWORK_MANAGEMENT_INCLUSION = 0x34;
 	public static int NETWORK_MANAGEMENT_PRIMARY = 0x54;
 	public static int ZWAVE_PLUS_INFO = 0x5E;
 	public static int MANUFACTURER_SPECIFIC = 0x72;
